kikimoraback/requirements.txt
amqp==5.2.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==25.1.0
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cramjam==2.8.3
cron-descriptor==1.4.5
dadata==24.10.0
Deprecated==1.2.15
distro==1.9.0
Django==5.1.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-imap-backend==0.2.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.2
idna==3.7
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.0
locust==2.32.5
Markdown==3.7
MarkupSafe==3.0.2
motor==3.6.0
msgpack==1.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.47
psutil==6.1.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
pymongo==4.9.2
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-snappy==0.7.2
pytz==2025.2
pyzmq==26.2.0
redis==4.5.5
requests==2.32.3
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
tomli==2.2.1
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn[standard]
var-dump==1.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yookassa==3.4.3
zope.event==5.0
zope.interface==7.2
